prototyped validators
