+++ src/components/input/input.ts
@@ -3,9 +3,9 @@ import template from 'bundle-text:./input.hbs';
 import './input.css';
 
 export interface InputProps {
-  onChange: (e: InputEvent) => void;
-  onFocus: (e: InputEvent) => void;
-  onBlur: (e: InputEvent) => void;
+  onChange?: (e: InputEvent) => void;
+  onFocus?: (e: InputEvent) => void;
+  onBlur?: (e: InputEvent) => void;
   type?: string;
   placeholder?: string;
   name?: string;
