+++ src/components/profile-edit-data/profileEditData.hbs
@@ -1,3 +1,3 @@
 <div class="profile-edit-data">
-  {{{Form inputs=inputs ref="formRef" formTitle=formTitle}}}
+  {{{Form inputs=inputs ref="formRef" formButton=formButton}}}
 </div>
