+++ src/components/profile-edit-password/profileEditPassword.hbs
@@ -1,4 +1,4 @@
 <div class="profile-edit-password">
-  {{{Form inputs=inputs ref="formRef" formTitle=formTitle}}}
+  {{{Form inputs=inputs ref="formRef" formButton=formButton}}}
 <div>
