+++ src/pages/Login/login.ts
@@ -1,7 +1,8 @@
 import Block from 'core/Block';
 import template from 'bundle-text:./login.hbs';
 import { InputProps } from 'components/input/input';
-import Validator from 'core/Validator';
+import ValidatorController from 'core/ValidatorController';
+
 interface LoginProps {
   formTitle: string;
   formName: string;
@@ -19,7 +20,7 @@ export class Login extends Block<LoginProps> {
   constructor(props: LoginProps) {
     super(props);
 
-    const validator = new Validator(this, props.inputs);
+    const validator = new ValidatorController(this, props.inputs);
     validator.init();
 
     this.eventBus.on(Block.EVENTS.FORM_SUBMIT, () => console.log(validator.errors));
