+++ src/pages/Profile/profile.ts
@@ -1,9 +1,43 @@
 import Block from '../../core/Block';
 import template from 'bundle-text:./profile.hbs';
+import { InputProps } from 'components/input/input';
+
 interface ProfileProps {
-  sideMenuExpanded: boolean;
-  feedEmpty: boolean;
-}
+  avatarInput: {
+    type: string;
+    name: string;
+    required: boolean;
+    id: string;
+    accept: string;
+  };
+  editPasswordForm: {
+    formTitle?: null;
+    formName?: string;
+    formButton: {
+      text: string;
+    };
+    inputs: InputProps[];
+    addon?: null;
+  };
+  editPersonalDataForm: {
+    formTitle?: null;
+    formName?: string;
+    formButton: {
+      text: string;
+    };
+    inputs: InputProps[];
+    addon?: null;
+  };
+  profileEditButtons: {
+    type: string;
+    name: string;
+    text: string;
+    link: string;
+  }[];
+  mainView?: boolean;
+  passwordView?: boolean;
+  dataView?: boolean;
+};
 
 
 export class Profile extends Block<ProfileProps> {
